backend/requirements.txt
# Framework principal
django==5.2.9

# API / REST
djangorestframework==3.16.1

# CORS
django-cors-headers==4.9.0

# HTTP requests (ZapSign)
requests==2.32.3

# Testes
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0

# Utilitários
python-dotenv==1.0.1
psycopg2-binary==2.9.10
python-dotenv

# Processamento Assíncrono
celery==5.4.0
redis==5.0.1

# IA / Machine Learning
google-generativeai==0.3.2  # Gemini API
openai==1.12.0  # OpenAI GPT API

# Processamento de PDF
PyPDF2==3.0.1  # Extração de texto de PDFs

# Autenticação
djangorestframework-simplejwt==5.3.1

# WebSockets
channels==4.1.0
channels_redis==4.2.0
daphne==4.1.2